Refactor generic helper methods
